Add House of 31 Game in Cpp
31" is a card game played by 2 or more people
players. Generally each player's goal is to collect a number from
playing cards whose sum is as close as possible to 31. If
the sum of one player is greater than 31 then he automatically exits
toy ("burned"). The winner is the player with the highest total that is
less than or equal to 31.